+++ spec/features/projects/viewing_another_users_or_unassigned_project_spec.rb
@@ -21,4 +21,12 @@ RSpec.feature "Viewing another user's or unassigned project" do
 
     expect(page).not_to have_button("Complete project")
   end
+
+  scenario "users do not see the submit button on tasks" do
+    project.task_list.tasks.each do |task|
+      visit conversions_voluntary_project_edit_task_path(project, task.class.identifier)
+
+      expect(page).not_to have_button("Save and return")
+    end
+  end
 end